temporary: fetch zmq from source because we're getting grief from SDK
...however, this ideally should be a build library that we obtain with the
toolchain itself. The main problem is that the toolchain provides libzmq as a
.so, however some system images in the field don't have that yet and would
require a firmware flash - as of now, I wasn't able to find a way to statically
link to zmq via the toolchain libraries. This is a temporary fix in the interim
until the fleet images are updated to ship with libzmq.so